+++ docs/extensions/engines/spark/rules.md
@@ -92,5 +92,5 @@ Kyuubi provides some configs to make these feature easy to use.
 | spark.sql.finalWriteStageExecutorMemory                             | fallback spark.executor.memory         | Specify the executor on heap memory request for final write stage. It would be passed to the RDD resource profile.                                                                                                                                                                                                                                   | 1.8.0 |
 | spark.sql.finalWriteStageExecutorMemoryOverhead                     | fallback spark.executor.memoryOverhead | Specify the executor memory overhead request for final write stage. It would be passed to the RDD resource profile.                                                                                                                                                                                                                                  | 1.8.0 |
 | spark.sql.finalWriteStageExecutorOffHeapMemory                      | NONE                                   | Specify the executor off heap memory request for final write stage. It would be passed to the RDD resource profile.                                                                                                                                                                                                                                  | 1.8.0 |
-| spark.sql.authz.scriptTransformation.enabled                        | false                                  | When false, script transformation is not allowed in authz.                                                                                                                                                                                                                                                                                           | 1.9.0 |
+| spark.sql.execution.scriptTransformation.enabled                    | true                                   | When false, script transformation is not allowed.                                                                                                                                                                                                                                                                                                    | 1.9.0 |
