+++ extensions/spark/kyuubi-extension-spark-3-5/src/main/scala/org/apache/kyuubi/sql/KyuubiSQLConf.scala
@@ -290,11 +290,10 @@ object KyuubiSQLConf {
       .intConf
       .createWithDefault(2000)
 
-  val AUTHZ_SCRIPT_TRANSFORMATION_ENABLED =
-    buildConf("spark.sql.authz.scriptTransformation.enabled")
-      .doc(
-        "When false, script transformation is not allowed in authz.")
+  val SCRIPT_TRANSFORMATION_ENABLED =
+    buildConf("spark.sql.execution.scriptTransformation.enabled")
+      .doc("When false, script transformation is not allowed.")
       .version("1.9.0")
       .booleanConf
-      .createWithDefault(false)
+      .createWithDefault(true)
 }
