+++ extensions/spark/kyuubi-extension-spark-common/src/main/scala/org/apache/kyuubi/sql/KyuubiSQLConf.scala
@@ -274,11 +274,10 @@ object KyuubiSQLConf {
       .stringConf
       .createOptional
 
-  val AUTHZ_SCRIPT_TRANSFORMATION_ENABLED =
-    buildConf("spark.sql.authz.scriptTransformation.enabled")
-      .doc(
-        "When false, script transformation is not allowed in authz.")
+  val SCRIPT_TRANSFORMATION_ENABLED =
+    buildConf("spark.sql.execution.scriptTransformation.enabled")
+      .doc("When false, script transformation is not allowed.")
       .version("1.9.0")
       .booleanConf
-      .createWithDefault(false)
+      .createWithDefault(true)
 }
